pyproject.toml
[tool.poetry]
name = "bookking"
version = "0.1.0"
description = ""
authors = ["Komarov Sergey <onubee@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
alembic = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
